refactor(lex): Pull out test mod
